mtKG-LLM/requirements.txt
openai>=1.40.0
groq>=0.9.0
networkx>=3.3
graspologic>=3.4.1
yacs>=0.1.8
Pillow>=10.4.0
numpy>=1.26.4

